examples/python/requirements.txt
Flask-RESTful==0.3.1
pymongo==2.7.2
